Move the approx thingy to the main module
