+++ pytest_flex_approx/__init__.py
@@ -1,0 +1,8 @@
+"""Flexible approx.
+
+Accepts dicts with a mixture of numbers and other things.
+"""
+
+from .flex_approx import flex_approx
+
+__all__ = ["flex_approx"]
